status function removal, preparation for universal login, other minor changes
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@9061 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/basics/baseparser.h
@@ -127,8 +127,8 @@ class CToken : public defaultcmp
 public:
 	short			id;
 	string<>		cLexeme;
-	unsigned short	line;
-	unsigned short	column;
+	unsigned int	line;
+	unsigned int	column;
 
 	CToken() : id(0),line(0),column(0)	{}
 	CToken(short i) : id(i), line(0),column(0)	{}
